+++ packages/runtime/adapters/eventstore-adapters/eventstore-postgresql/src/set-replication-status.ts
@@ -8,9 +8,17 @@ import initReplicationStateTable from './init-replication-state-table'
 
 const setReplicationStatus = async (
   pool: AdapterPool,
-  status: ReplicationStatus,
-  statusData?: ReplicationState['statusData'],
-  lastEvent?: OldEvent
+  {
+    status,
+    statusData,
+    lastEvent,
+    iterator,
+  }: {
+    status: ReplicationStatus
+    statusData?: ReplicationState['statusData']
+    lastEvent?: OldEvent
+    iterator?: ReplicationState['iterator']
+  }
 ): Promise<void> => {
   const { executeStatement, escapeId, escape, databaseName } = pool
 
@@ -27,6 +35,13 @@ const setReplicationStatus = async (
       ${
         lastEvent != null
           ? `, "SuccessEvent" = ${escape(JSON.stringify(lastEvent))}`
+          : ``
+      }
+      ${
+        iterator !== undefined
+          ? `, "Iterator" = ${
+              iterator != null ? escape(JSON.stringify(iterator)) : 'NULL'
+            }`
           : ``
       }`
   )